packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-qa-suites"
version = "0.1.0"
description = "Test suite definitions, test cases, and assertions for validating clawspace toolboxes"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["qa-suites","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-qa-suites"
Repository = "https://github.com/mxn2020/minions-qa-suites"

[tool.hatch.build.targets.wheel]
packages = ["minions_qa_suites"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
